backend/utils/snowflake agent/requirements.txt
snowflake-connector-python
yfinance
python-dotenv